+++ CHANGELOG.md
@@ -10,6 +10,11 @@ changes.
 
 ## [0.13.0] - UNRELEASED
 
+- Added retries for local cardano-node queries. We witnessed these queries to
+  fail in case of a rollback so we ignore errors and retry for 5 times before
+  letting the query fail.
+
+
 - **BREAKING** Changes to `hydra-plutus` scripts.
 
 - Add option to draft a commit tx using inline datums.